requirements.txt
torch 
torchvision 
opencv-python 
scikit-learn 
segment-anything 
pillow 
pyyaml